Rename interfaces.ts to types.ts. Add CategoryTypes
