+++ ui/src/screens/Posts/components/PostDetail.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import styled from "styled-components";
 
-import { PostInterface } from "../../../interfaces";
+import { PostInterface } from "../../../types";
 
 const Container = styled.div`
   background-color: white;
